fix(validation): reject placeholder GitHub usernames at write side
Mirrors the runtime auto-repair guard in pr-monitor (placeholder values
like "example-user" auto-corrected from the authenticated viewer) by
rejecting the same values at the entry-point validator. Without this,
init / setup --set username= / the MCP config tool would persist
"example-user" or "your-username" copied from documentation, leaving
the user with a "Showing partial data" banner on the dashboard until
the next fetch silently repaired the config.

Extracts PLACEHOLDER_USERNAMES + isPlaceholderUsername to a small
dedicated module so commands/validation.ts can import the predicate
without pulling in pr-monitor's octokit / github-stats dependency tree.

Adds a parameterized test covering the rejection (case-insensitive)
and renames a now-misleading literal in tools.test.ts (mocked schema
test that incidentally used "example-user" as a sample value).
